fix(pa-scops-oag): address Copilot review — i64 move guard, graceful test teardown
- move_(): compare position against max_step in i64. max_step is u32, so the
  old `as i32` cast wrapped negative for max_step > i32::MAX and rejected every
  move. New BDD scenario pins the boundary (max step 4294967295, move to
  2147483647) using only existing steps.
- test_lib.rs: spawn_server hands back a oneshot stop sender wired into the
  shutdown future; tests stop the server gracefully and await the join handle
  instead of abort()ing, so BoundServer::start()'s teardown (transport
  shutdown) actually runs — no leaked supervisor/poll tasks across the test
  process, and the teardown path gains test coverage.

The get_status retry suggestion was declined: build() binds the listener
before spawn_server returns, so a connect always lands in the kernel accept
backlog — there is no readiness race to retry around.

+++ services/pa-scops-oag/src/focuser_device.rs
@@ -203,7 +203,9 @@ impl Focuser for ScopsFocuserDevice {
     async fn move_(&self, position: i32) -> ASCOMResult<()> {
         ensure_connected!(self);
 
-        if position < 0 || position > self.config.max_step as i32 {
+        // Compare in i64: `max_step` is u32 and may exceed `i32::MAX`, where a
+        // `as i32` cast would wrap negative and reject every move.
+        if position < 0 || i64::from(position) > i64::from(self.config.max_step) {
             return Err(ASCOMError::new(
                 ASCOMErrorCode::INVALID_VALUE,
                 format!(
